examples/with_bridge/requirements.txt
azure-functions
azure-functions-openapi[bridge]
